# container/spec/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fire Pocker API
  version: 0.1.0
  description: >
    REST API definition for Fire Pocker. All endpoints are served from the
    Next.js App Router running inside the Fargate task.
servers:
  - url: https://api.example.com
    description: Production (placeholder)
  - url: https://stg.api.example.com
    description: Staging (placeholder)
...
components:
  securitySchemes:
    SessionSecretHeader:
      type: apiKey
      in: header
      name: x-session-secret
      description: Shared secret used by facilitators to create sessions and fetch PBIs.
    JoinToken:
      type: http
      scheme: bearer
      bearerFormat: HMAC
      description: Short lived token issued per session participant.
  schemas:
    PBI:
      type: object
      properties:
        id:
          type: string
          example: 12345678-abcd-ef00-1234-56789abcdef0
        title:
          type: string
        status:
          type: string
          example: In Progress
        storyPoint:
          type: integer
          nullable: true
        assignee:
          type: string
          nullable: true
        epic:
          type: string
          nullable: true
        lastEstimatedAt:
          type: string
          format: date-time
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
paths:
  /api/pbis:
    get:
      tags: [PBI]
      summary: Notion の PBI一覧DB から見積もり対象候補の一覧を取得する
      security:
        - SessionSecretHeader: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
          required: false
        - name: search
          in: query
          schema:
            type: string
          required: false
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PBI'
                  nextCursor:
                    type: string
                    nullable: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
...